feat(backtest): persist runs, trades, optimisations and trials
Four tables behind one repository module with a draft → queued → running →
done/failed lifecycle, results saved atomically with their trades, and deletes
cascading through optimisations and trials. Guarded Alembic migration so a DB
that already got the tables from create_all is a no-op. Round-trip test on
in-memory SQLite.

+++ strategies/parameters.py
@@ -82,7 +82,7 @@ class DecimalParameter(Parameter):
         return len(text.split(".")[1]) if "." in text else 0
 
     def grid(self) -> list[float]:
-        count = int(round((self.high - self.low) / self.step))
+        count = round((self.high - self.low) / self.step)
         return [round(self.low + i * self.step, self.decimals) for i in range(count + 1)]
